refactor: replace to axios
